--- src/auth/interfaces/authenticated-request.interface.ts
@@ -1,6 +1,0 @@
-import { Request } from 'express';
-import { AuthenticatedUser } from './authenticated-user';
-
-export interface AuthenticatedRequest extends Request {
-  user: AuthenticatedUser;
-}
