+++ src/auth/interfaces/user-authenticated-request.ts
@@ -1,0 +1,5 @@
+import { UserAuthenticated } from './user-authenticated';
+
+export interface UserAuthenticatedRequest extends Request {
+  user: UserAuthenticated;
+}
