+++ src/auth/interfaces/user-authenticated.ts
@@ -1,6 +1,6 @@
 import { UsuarioStatus } from 'src/entities/user.entity';
 
-export interface AuthenticatedUser {
+export interface UserAuthenticated {
   id: number;
   email_institucional: string;
   status_usuario: UsuarioStatus;
